requirements.txt
asgi_redis
channels==3.0.5
channels_redis
Django
django-cors-headers
djangorestframework
djoser
mysqlclient
python-dotenv